recast float32 for plotting

+++ invldm/utils/visualisation.py
@@ -9,6 +9,6 @@ def visualise_samples(x, scale=False):
     if scale:
         x = scale2range(x, [0, 1])
     fig = plt.figure()
-    grid = make_grid(x).permute(1, 2, 0).cpu()
+    grid = make_grid(x.float()).permute(1, 2, 0).cpu()
     plt.imshow(grid)
     return fig
